Refactor debug page path
