+++ src/lib/ui/debug/protected-viewer.tsx
